+++ internal/pipeline/pipeline.go
@@ -17,34 +17,14 @@
 package pipeline
 
 import (
-	"fmt"
-
 	"github.com/GoogleContainerTools/kpt/pkg/api/kptfile/v1alpha2"
 	"sigs.k8s.io/kustomize/kyaml/kio"
-	"sigs.k8s.io/kustomize/kyaml/yaml"
 )
 
 const (
 	sourceAllSubPkgs string = "./*"
 	sourceCurrentPkg string = "."
 )
-
-// ValidatePipeline will validate all fields in the Pipeline
-// 'generators', 'transformers' and 'validators' share same schema and
-// they are valid if all functions in them are ALL valid.
-func ValidatePipeline(p *v1alpha2.Pipeline) error {
-	fns := []v1alpha2.Function{}
-	fns = append(fns, p.Mutators...)
-	fns = append(fns, p.Validators...)
-	for i := range fns {
-		f := fns[i]
-		err := validateFunction(&f)
-		if err != nil {
-			return fmt.Errorf("function %q is invalid: %w", f.Image, err)
-		}
-	}
-	return nil
-}
 
 // fnChain returns a slice of function runners from the
 // functions and configs defined in pipeline.
@@ -63,18 +43,4 @@ func fnChain(p *v1alpha2.Pipeline, pkgPath string) ([]kio.Filter, error) {
 		runners = append(runners, r)
 	}
 	return runners, nil
-}
-
-// fromBytes returns a Pipeline parsed from bytes
-func fromBytes(b []byte) (*v1alpha2.Pipeline, error) {
-	p := &v1alpha2.Pipeline{}
-	err := yaml.Unmarshal(b, p)
-	if err != nil {
-		return nil, fmt.Errorf("failed to construct pipeline from bytes: %w", err)
-	}
-	return p, nil
-}
-
-func fromString(s string) (*v1alpha2.Pipeline, error) {
-	return fromBytes([]byte(s))
 }
