+++ internal/pipeline/pipeline_function_test.go
@@ -118,164 +118,3 @@ data: {foo: bar}
 		})
 	}
 }
-
-func TestValidateFunctionName(t *testing.T) {
-	type input struct {
-		Name  string
-		Valid bool
-	}
-	inputs := []input{
-		{
-			"gcr.io/kpt-functions/generate-folders",
-			true,
-		},
-		{
-			"patch-strategic-merge",
-			true,
-		},
-		{
-			"a.b.c:1234/foo/bar/generate-folders",
-			true,
-		},
-		{
-			"ab-.b/c",
-			false,
-		},
-		{
-			"a/a/",
-			false,
-		},
-		{
-			"a//a/a",
-			false,
-		},
-		{
-			"example.com/.dots/myimage",
-			false,
-		},
-		{
-			"registry.io/foo/project--id.module--name.ver---sion--name",
-			true,
-		},
-		{
-			"Foo/FarB",
-			false,
-		},
-	}
-
-	for _, n := range inputs {
-		n := n
-		t.Run(n.Name, func(t *testing.T) {
-			err := validateFunctionName(n.Name)
-			if n.Valid && err != nil {
-				t.Fatalf("function name %s should be valid", n.Name)
-			}
-			if !n.Valid && err == nil {
-				t.Fatalf("function name %s should not be valid", n.Name)
-			}
-		})
-
-	}
-}
-
-func TestValidatePath(t *testing.T) {
-	type input struct {
-		Path  string
-		Valid bool
-	}
-
-	cases := []input{
-		{
-			"a/b/c",
-			true,
-		},
-		{
-			"a/b/",
-			true,
-		},
-		{
-			"/a/b",
-			false,
-		},
-		{
-			"./a",
-			true,
-		},
-		{
-			"./a/.../b",
-			true,
-		},
-		{
-			".",
-			true,
-		},
-		{
-			"a\\b",
-			false,
-		},
-		{
-			"a\b",
-			false,
-		},
-		{
-			"a\v",
-			false,
-		},
-		{
-			"a:\\b\\c",
-			false,
-		},
-		{
-			"../a/../b",
-			true,
-		},
-		{
-			"a//b",
-			true,
-		},
-		{
-			"a/b/.",
-			true,
-		},
-		{
-			"a/*/b",
-			false,
-		},
-		{
-			"./*",
-			true,
-		},
-		{
-			"a/b\\c",
-			false,
-		},
-		{
-			"././././",
-			true,
-		},
-		{
-			"./!&^%$/#(@)/_-=+|<;>?:'\"/'`",
-			true,
-		},
-		{
-			"",
-			false,
-		},
-		{
-			"\t \n",
-			false,
-		},
-		{
-			"*",
-			false,
-		},
-	}
-
-	for _, c := range cases {
-		ret := validatePath(c.Path)
-		if (ret == nil) != c.Valid {
-			t.Fatalf("returned value for path %s should be %t, got %t",
-				c.Path, c.Valid, (ret == nil))
-		}
-	}
-}
